Add files via upload
uploading myEvery function. Works on the same principles as the mySome but with the conditions flipped. Since every item in the array must work with the function, it will stop as soon as it sees one that does not and return false. if it goes through all of them, with out seeing any, it returns true.