Add files via upload
adding myReduces. Since reduce has some specific circumstances, did two different versions of the for loop. If an initial value is passed, the for loop is done on all the values in the array. If there isnt one passed, the inital value is set to the first value of the array and then allows for the whole array to pass.