Add files via upload
Adding my push and my last index of. Last index of is the same as index of but with the loop starting at the back end. It returns the value as soon as it finds it and returns -1 if not found. Mypush takes the arguments and loops them to pushing them to the end of the array as well as returning the length of new array.

Several small bug fixes have also been done that were found after running some tests.